use Backend interface for shards and allow to pass a builder function
