+++ krakendrate.go
@@ -4,7 +4,6 @@ package krakendrate
 import (
 	"context"
 	"errors"
-	"sync"
 	"time"
 )
 
@@ -38,9 +37,13 @@ type Backend interface {
 	Store(string, interface{}) error
 }
 
+// BackendBuilder is the type for a function that can build a Backend.
+// Is is used by the ShardedMemoryBackend to create several backends / shards.
+type BackendBuilder func(ctx context.Context, ttl time.Duration) Backend
+
 // ShardedMemoryBackend is a memory backend shardering the data in order to avoid mutex contention
 type ShardedMemoryBackend struct {
-	shards []*MemoryBackend
+	shards []Backend
 	total  uint64
 	hasher Hasher
 }
@@ -52,14 +55,20 @@ func DefaultShardedMemoryBackend(ctx context.Context) *ShardedMemoryBackend {
 
 // NewShardedMemoryBackend returns a ShardedMemoryBackend with 'shards' shards
 func NewShardedMemoryBackend(ctx context.Context, shards uint64, ttl time.Duration, h Hasher) *ShardedMemoryBackend {
+	return NewShardedBackend(ctx, shards, ttl, h, MemoryBackendBuilder)
+}
+
+func NewShardedBackend(ctx context.Context, shards uint64, ttl time.Duration, h Hasher,
+	backendBuilder BackendBuilder) *ShardedMemoryBackend {
+
 	b := &ShardedMemoryBackend{
-		shards: make([]*MemoryBackend, shards),
+		shards: make([]Backend, shards),
 		total:  shards,
 		hasher: h,
 	}
 	var i uint64
 	for i = 0; i < shards; i++ {
-		b.shards[i] = NewMemoryBackend(ctx, ttl)
+		b.shards[i] = backendBuilder(ctx, ttl)
 	}
 	return b
 }
@@ -78,6 +87,12 @@ func (b *ShardedMemoryBackend) Store(key string, v interface{}) error {
 	return b.shards[b.shard(key)].Store(key, v)
 }
 
+/*
+
+This function looks like is not called from anywhere and allows us to
+avoid using the specific MemoryBackend type and just use the
+BackendInterface.
+
 func (b *ShardedMemoryBackend) del(key ...string) {
 	buckets := map[uint64][]string{}
 	for _, k := range key {
@@ -95,98 +110,4 @@ func (b *ShardedMemoryBackend) del(key ...string) {
 		b.shards[s].del(ks...)
 	}
 }
-
-func NewMemoryBackend(ctx context.Context, ttl time.Duration) *MemoryBackend {
-	m := &MemoryBackend{
-		data:       map[string]interface{}{},
-		lastAccess: map[string]time.Time{},
-		mu:         new(sync.RWMutex),
-	}
-
-	go m.manageEvictions(ctx, ttl)
-
-	return m
-}
-
-// MemoryBackend implements the backend interface by wrapping a sync.Map
-type MemoryBackend struct {
-	data       map[string]interface{}
-	lastAccess map[string]time.Time
-	mu         *sync.RWMutex
-}
-
-func (m *MemoryBackend) manageEvictions(ctx context.Context, ttl time.Duration) {
-	t := time.NewTicker(ttl)
-	for {
-		var keysToDel []string
-
-		select {
-		case <-ctx.Done():
-			t.Stop()
-			return
-		case now := <-t.C:
-			m.mu.RLock()
-			for k, v := range m.lastAccess {
-				if v.Add(ttl).Before(now) {
-					keysToDel = append(keysToDel, k)
-				}
-			}
-			m.mu.RUnlock()
-		}
-
-		m.del(keysToDel...)
-	}
-}
-
-// Load implements the Backend interface
-func (m *MemoryBackend) Load(key string, f func() interface{}) interface{} {
-	m.mu.RLock()
-	v, ok := m.data[key]
-	m.mu.RUnlock()
-
-	n := now()
-
-	if ok {
-		go func(t time.Time) {
-			m.mu.Lock()
-			if t0, ok := m.lastAccess[key]; !ok || t.After(t0) {
-				m.lastAccess[key] = t
-			}
-			m.mu.Unlock()
-		}(n)
-
-		return v
-	}
-
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	v, ok = m.data[key]
-	if ok {
-		return v
-	}
-
-	v = f()
-	m.lastAccess[key] = n
-	m.data[key] = v
-
-	return v
-}
-
-// Store implements the Backend interface
-func (m *MemoryBackend) Store(key string, v interface{}) error {
-	m.mu.Lock()
-	m.lastAccess[key] = now()
-	m.data[key] = v
-	m.mu.Unlock()
-	return nil
-}
-
-func (m *MemoryBackend) del(key ...string) {
-	m.mu.Lock()
-	for _, k := range key {
-		delete(m.data, k)
-		delete(m.lastAccess, k)
-	}
-	m.mu.Unlock()
-}
+*/
